fix bugs & add run scripts

+++ website/models.py
@@ -5,7 +5,7 @@ from sqlalchemy.sql import func
 
 class Note(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    amount = db.Column(db.Integer, default=0)
+    amount = db.Column(db.Float, default=0)
     type = db.Column(db.String(150), nullable=False)
     comment = db.Column(db.String(1500), default="No comment")
     date = db.Column(db.DateTime(timezone=True), default=func.now())
